vaults move: ../data_sources/[dsId]/tables/[tId]/index
